Trying to make UploadProduct custom hook
